try and make work with travis and python3.4
